Redone image/pixel convertion, added runtime endiannes check, updated for testing

+++ BubbleFighter/BubbleFighter.NativeActivity/core/DataStream.h
@@ -5,7 +5,7 @@
 
 namespace core
 {
-	typedef long int data_size;
+	typedef int data_size;
 
 	class DataStream
 	{
